begin factoring output stream logic into separate widget for reuse
