+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/ChunkConsolePage.java
@@ -1,0 +1,37 @@
+/*
+ * ChunkConsolePage.java
+ *
+ * Copyright (C) 2009-16 by RStudio, Inc.
+ *
+ * Unless you have received this program directly from RStudio pursuant
+ * to the terms of a commercial license agreement with RStudio, then
+ * this program is licensed to you under the terms of version 3 of the
+ * GNU Affero General Public License. This program is distributed WITHOUT
+ * ANY EXPRESS OR IMPLIED WARRANTY, INCLUDING THOSE OF NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE. Please refer to the
+ * AGPL (http://www.gnu.org/licenses/agpl-3.0.txt) for more details.
+ *
+ */
+package org.rstudio.studio.client.workbench.views.source.editors.text;
+
+import com.google.gwt.user.client.ui.HTML;
+import com.google.gwt.user.client.ui.Widget;
+
+public class ChunkConsolePage implements ChunkOutputPage
+{
+   @Override
+   public Widget thumbnailWidget()
+   {
+      return new HTML("Console");
+   }
+
+   @Override
+   public Widget contentWidget()
+   {
+      return null;
+   }
+   
+   public final static int CONSOLE_INPUT  = 0;
+   public final static int CONSOLE_OUTPUT = 1;
+   public final static int CONSOLE_ERROR  = 2;
+}
